Update multiexp.py
This is a PSF class that can take stars from multiple exposures are input, then create individual PSFs from each exposure. PSFs are computed in series rather than parallel to each other, so this code takes longer to run, though it is able to plot output from each PSF on the same graph.